refactor: expirationDate를 custom hook으로 관리
